+++ js/data.js
@@ -1,0 +1,64 @@
+const COUNT_ADS = 10;
+const MIN_PRICE = 100;
+const MAX_PRICE = 1000;
+const MAX_ROOMS = 10;
+const MAX_GUESTS = 10;
+const MIN_LOCATION_LAT = 35.65;
+const MAX_LOCATION_LAT = 35.7;
+const MIN_LOCATION_LNG = 139.7;
+const MAX_LOCATION_LNG = 139.8;
+const DECIMAL = 5;
+
+const PHOTOS_AD = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const FEATURES_AD = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const TIMES_RENTAL = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const TIPES_AD = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const TITLES_AD = [
+  'Cheap',
+  'Wonderful',
+  'Cool',
+  'Great',
+  'Mind-blowing',
+];
+
+export {
+  COUNT_ADS,
+  MIN_PRICE,
+  MAX_PRICE,
+  MAX_ROOMS,
+  MAX_GUESTS,
+  MIN_LOCATION_LAT,
+  MAX_LOCATION_LAT,
+  MIN_LOCATION_LNG,
+  MAX_LOCATION_LNG,
+  DECIMAL,
+  PHOTOS_AD,
+  FEATURES_AD,
+  TIMES_RENTAL,
+  TIPES_AD, TITLES_AD
+};
